Drop Algebra.Graph.AdjacencyMap.Internal
